requirements.txt
pyaudio
pyttsx3
soundfile
noisereduce
webrtcvad
librosa
numpy
googletrans
google.cloud
sounddevice
resemblyzer